split funcs in storage.go
